Move ViewManager out of core
