+++ cmp/viewmanager/ViewInfo.ts
