+++ cmp/viewmanager/ViewManagerModel.ts
@@ -20,7 +20,7 @@ import {
     Thunkable,
     XH
 } from '@xh/hoist/core';
-import type {ViewManagerProvider} from '../provider/ViewManagerProvider';
+import type {ViewManagerProvider} from '@xh/hoist/core';
 import {genDisplayName} from '@xh/hoist/data';
 import {fmtDateTime} from '@xh/hoist/format';
 import {action, makeObservable, observable, runInAction, when} from '@xh/hoist/mobx';
