+++ cmp/viewmanager/index.ts
