+++ core/persist/provider/ViewManagerProvider.ts
@@ -8,7 +8,7 @@
 import {throwIf} from '@xh/hoist/utils/js';
 import {pull} from 'lodash';
 import {PersistenceProvider, PersistenceProviderConfig} from '../PersistenceProvider';
-import type {ViewManagerModel} from '../viewmanager/ViewManagerModel';
+import type {ViewManagerModel} from '@xh/hoist/cmp/viewmanager/ViewManagerModel';
 
 export class ViewManagerProvider<S> extends PersistenceProvider<S> {
     readonly viewManagerModel: ViewManagerModel;
